+++ cmd/faman/main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"os"
+
+	"github.com/faman-project/faman/internal/app"
+)
+
+func main() {
+	if err := app.Execute(); err != nil {
+		os.Exit(1)
+	}
+}